benefit

+++ src/main/java/com/tiep/demoapus/service/RecruitmentChannelService.java
@@ -6,7 +6,7 @@ import com.tiep.demoapus.dto.response.RecruitmentChannelResponseDTO;
 
 public interface RecruitmentChannelService {
     PageableResponse<RecruitmentChannelResponseDTO> getAllRecruitmentChannels(int page, int size, String sort, String search);
-    RecruitmentChannelResponseDTO getRecruitmentChannel(int id);
+    RecruitmentChannelResponseDTO getRecruitmentChannel(Long id);
     RecruitmentChannelResponseDTO addRecruitmentChannel(RecruitmentChannelRequestDTO requestDTO);
     RecruitmentChannelResponseDTO updateRecruitmentChannel(RecruitmentChannelRequestDTO requestDTO);
     void deleteRecruitmentChannel(Long id);
